refactor: attachment 패키지를 계층형 구조로 정리
Attachment 는 JPA 엔티티가 아니라 도메인 간 첨부 정보를 주고받는 전달용 값 타입이다.
리포지토리도 없다. 실제 영속화는 각 도메인의 XxxFileUpload 엔티티가 한다.
따라서 Port/Adapter 제거나 엔티티 통합 없이 패키지 이동만 한다.

  application  -> service
  presentation -> controller
  presentation/dto -> dto

AttachmentService 인터페이스도 제거했다. 구현체가 하나뿐이고 default 메서드도 없어
계약만 있는 껍데기였다.

  attachment  8파일 788줄 -> 7파일

See also: #15
