+++ src/main/java/com/edu/edumeet/board/application/BoardServiceImpl.java
@@ -7,8 +7,8 @@ import com.edu.edumeet.board.presentation.dto.*;
 import com.edu.edumeet.reply.application.ReplyRepository;
 import com.edu.edumeet.s3.util.S3Uploader;
 import com.edu.edumeet.attachment.domain.Attachment;
-import com.edu.edumeet.attachment.presentation.dto.AttachmentAdapter;
-import com.edu.edumeet.attachment.presentation.AttachmentService;
+import com.edu.edumeet.attachment.dto.AttachmentAdapter;
+import com.edu.edumeet.attachment.service.AttachmentService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 import org.modelmapper.ModelMapper;
