+++ src/main/java/com/edu/edumeet/board/presentation/BoardService.java
@@ -1,7 +1,7 @@
 package com.edu.edumeet.board.presentation;
 
 
-import com.edu.edumeet.attachment.presentation.dto.AttachmentDTO;
+import com.edu.edumeet.attachment.dto.AttachmentDTO;
 import com.edu.edumeet.board.domain.Board;
 import com.edu.edumeet.board.domain.BoardType;
 import com.edu.edumeet.board.presentation.dto.*;
