+++ src/main/java/com/edu/edumeet/board/presentation/dto/BoardDTO.java
@@ -1,6 +1,6 @@
 package com.edu.edumeet.board.presentation.dto;
 
-import com.edu.edumeet.attachment.presentation.dto.AttachmentDTO;
+import com.edu.edumeet.attachment.dto.AttachmentDTO;
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.Size;
