+++ src/main/java/com/edu/edumeet/classroom/controller/ClassThumbnailController.java
@@ -1,7 +1,7 @@
 package com.edu.edumeet.classroom.controller;
 
 import com.edu.edumeet.classroom.service.ClassThumbnailService;
-import com.edu.edumeet.attachment.presentation.dto.AttachmentUploadDTO;
+import com.edu.edumeet.attachment.dto.AttachmentUploadDTO;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 import org.springframework.http.ResponseEntity;
