+++ src/main/java/com/edu/edumeet/homework/dto/AssignmentCreateDTO.java
@@ -1,7 +1,7 @@
 package com.edu.edumeet.homework.dto;
 
 import com.edu.edumeet.attachment.domain.Attachment;
-import com.edu.edumeet.attachment.presentation.dto.AttachmentDTO;
+import com.edu.edumeet.attachment.dto.AttachmentDTO;
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
