+++ src/test/java/com/edu/edumeet/unit/homework/application/AssignmentServiceImplTest.java
@@ -11,8 +11,8 @@ import com.edu.edumeet.homework.domain.Assignment;
 import com.edu.edumeet.homework.service.AssignmentService;
 import com.edu.edumeet.homework.dto.AssignmentCreateDTO;
 import com.edu.edumeet.homework.dto.AssignmentDTO;
-import com.edu.edumeet.attachment.presentation.dto.AttachmentAdapter;
-import com.edu.edumeet.attachment.presentation.dto.AttachmentDTO;
+import com.edu.edumeet.attachment.dto.AttachmentAdapter;
+import com.edu.edumeet.attachment.dto.AttachmentDTO;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
